and block matching works too!

+++ src/bmMatch.h
@@ -11,7 +11,7 @@ namespace stereo {
 		virtual void match(img::Img<uint8_t> & left, img::Img<uint8_t> & right,
 			img::Img<uint16_t> & disp) override;
 	private:
-		std::vector<uint32_t> edgeLeft, edgeRight;
-		std::vector<uint16_t> costs;
+		std::vector<int16_t> edgeLeft, edgeRight;
+		std::vector<float> costs;
 	};
 }
